refactor: swagger 및 rest API 명세 수정
