--- certbot/conf/accounts/acme-v02.api.letsencrypt.org/directory/053f50e92556dc2ae6fb1958ad04b4cb/meta.json
@@ -1,1 +1,0 @@
-{"creation_dt": "2024-01-16T08:39:49Z", "creation_host": "1b0b13fa81e7"}
