--- certbot/conf/accounts/acme-v02.api.letsencrypt.org/directory/053f50e92556dc2ae6fb1958ad04b4cb/regr.json
@@ -1,1 +1,0 @@
-{"body": {}, "uri": "https://acme-v02.api.letsencrypt.org/acme/acct/1519022676"}
